+++ packages/api/src/index.ts
@@ -4,3 +4,4 @@ export * from './definitions';
 export * from './objects';
 export * from './organizations';
 export * from './provenance-events';
+export * from './provenance-events-searcher';
