+++ packages/api/src/provenance-events-searcher/index.ts
@@ -1,0 +1,2 @@
+export * from './searcher-constituents-datahub';
+export * from './searcher-constituents-wikidata';
