--- packages/api/src/provenance-events/definitions.ts
@@ -1,5 +1,0 @@
-import {Thing} from '../definitions';
-
-export type SearchResult = {
-  things: Thing[];
-};
